Add code duplication check
